fixing auto-inc id

+++ AppDbContext/Models/Attribute.cs
@@ -14,10 +14,10 @@ namespace AppDbContext.Models
             CategoryAttribute = new HashSet<CategoryAttribute>();
         }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public long ValueTypeId { get; set; }
+        public int ValueTypeId { get; set; }
 
         public virtual ValueType ValueType { get; set; }
         public virtual ICollection<AttributeProductValue> AttributeProductValue { get; set; }
