+++ AppDbContext/Models/AttributeProductValue.cs
@@ -6,9 +6,9 @@ namespace AppDbContext.Models
 {
     public partial class AttributeProductValue
     {
-        public long Id { get; set; }
-        public long AttributeId { get; set; }
-        public long ProductId { get; set; }
+        public int Id { get; set; }
+        public int AttributeId { get; set; }
+        public int ProductId { get; set; }
         public string Value { get; set; }
 
         public virtual Attribute Attribute { get; set; }
